Move canned data to file
